requirements.txt
click==7.1.2
ipykernel==5.3.4
joblib==0.16.0
matplotlib==3.3.1
notebook==6.1.3
pandas==1.1.
pytest==6.0.1
PyYAML==5.3.1
rich==5.2.0
tqdm==4.48.2
